pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-agent"
version = "0.1.0"
description = "A clean, minimal AI agent framework"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "zhipuai>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["simple_agent*"]
